requirements.txt
python-telegram-bot==20.2
pypr0==0.2.7